% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_connection.R
\name{bed_connection}
\alias{bed_connection}
\title{Create a connection to IOBED's beds}
\usage{
bed_connection(port, buffersize = 2^21)
}
\arguments{
\item{port}{(chr) the serial port used for the connection.}

\item{buffersize}{(int, default 2^21, i.e. ~2 MB) number of bytes
to dedicate to the buffer.}
}
\value{
a connection
}
\description{
Setup a connection to the bed serial output with all the
specifications needed for the type of its output stream.
}
\note{
The number of bytes used is set to approx 2 MB because the bed
stream a line of information of less than 100 characters (about
50-60), collecting sensors signals at 5Hz. So, at 1 Byte each
character, it sum up to 0.5 kB/s. 2 MB of buffer should be
sufficient to collect 4000 seconds of data (~ 1 hour), that should
be both a sufficient and reasonable amount of time for a single
session of data collection.
}
\examples{
\dontrun{
  # connect the serial cable to the computer and
  con <- bed_connection() # default port is "COM3"

  port_used <- "COM1"
  bed_connection(port = port_used)
}
}
