% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit2datacsv.R
\name{fit2datacsv}
\alias{fit2datacsv}
\title{Convert FIT-data to CSV}
\usage{
fit2datacsv(fit_path, fit_jar_path = "C:\\\\FitSDK\\\\java\\\\FitCSVTool.jar")
}
\arguments{
\item{fit_path}{(chr) path to the source FIT data file}

\item{fit_jar_path}{(chr, default 'C:\\FitSDK\\java\\FitCSVTool.jar')
path to the jar Garmin converter}
}
\value{
path to the converted csv
}
\description{
Convert FIT-data to CSV
}
