% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iobed.watch-package.R
\docType{package}
\name{iobed.watch-package}
\alias{iobed.watch}
\alias{iobed.watch-package}
\title{iobed.watch: IOBED functions for bed's raw-signals managment at UBEP}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

An utility package for the management of raw-signals from the Malvestio's smart-bed at UBEP within the IOBED project.
}
\author{
\strong{Maintainer}: Corrado Lanera \email{corrado.lanera@ubep.unipd.it} (\href{https://orcid.org/0000-0002-0520-7428}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Carlotta Borghini \email{carlotta.borghini@ubep,unipd.it}
  \item Danila Azzolina \email{danila.azzolina@ubep.unipd.it} (\href{https://orcid.org/0000-0002-8185-5742}{ORCID})
  \item Alessandra Casamento \email{alessandra.casamento@ubep.unipd.it} (\href{https://orcid.org/0000-0001-6156-9020}{ORCID})
  \item Dario Gregori \email{dario.gregori@ubep,unipd.it} (\href{https://orcid.org/0000-0001-7906-0580}{ORCID}) [copyright holder]
}

}
\keyword{internal}
