% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fit.R
\name{load_fit}
\alias{load_fit}
\title{Load FIT file}
\usage{
load_fit(fit_source_path, output_dir, overwrite = FALSE)
}
\arguments{
\item{fit_source_path}{(chr) file path for the input FIT file.}

\item{output_dir}{(chr) path of the directory into which copy the
\code{fit_source_path}.}

\item{overwrite}{(lgl, NULL) do you want to overwrite potential
already existing file?}
}
\value{
(chr) output file path
}
\description{
It copies a FIT file (\code{.fit}) from a location (e.g., the watch
folder) to another location (e.g., person data folder).
}
