% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_iobed_stream.R
\name{tidy_iobed_stream}
\alias{tidy_iobed_stream}
\title{Tidy version of IOBED stream}
\usage{
tidy_iobed_stream(stream)
}
\arguments{
\item{stream}{(chr) a stream from the IOBED connection}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get a stream pulled by \link{pull_bed_stream} and tidy it in a tibble
}
\note{
If the first row is uncompleted it is ignored and a warning
is signaled (this should not happen!), on the other hand, if the
last row is not completed (almost always!) it is silently ignored.
}
\examples{
sample_stream <- c(
  "2", "9", "6", " ", " ", " ", " ", "1", "9", "5", " ", " ",
  " ", "2", "2", "9", " ", " ", " ", " ", "2", "7", "7", " ", " ",
  " ", " ", "3", "8", " ", " ", " ", " ", " ", "1", "6", ":", "4",
  "2", ":", "3", "8", "\n", "\n", "", "2", "9", "6", " ", " ",
  " ", " ", "1", "9", "5", " ", " ", " ", "2", "2", "9", " ", " ",
  " ", " ", "2", "7", "7", " ", " ", " ", " ", "3", "8", " ", " ",
  " ", " ", " ", "1", "6", ":", "4", "2", ":", "3", "9", "\n",
  "\n", "", "2", "9", "6", " ", " ", " ", " ", "1", "9", "5"
)

tidy_iobed_stream(sample_stream)

}
