% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fit_data.R
\name{parse_fit_data}
\alias{parse_fit_data}
\title{Parse FIT data}
\usage{
parse_fit_data(fit_data_csv_path, dir_path = NULL, prefix_name = NULL)
}
\arguments{
\item{fit_data_csv_path}{(chr) file path the the \verb{<name>_data.csv}
file as output from the function \code{fit2datacsv}}

\item{dir_path}{(chr, NULL) if not NULL (default) the dir path in
which to write the resulting tibble as .rds file}

\item{prefix_name}{(chr, NULL) optional prefix to add to the filename}
}
\value{
a \link[tibble:tibble-package]{tibble} with the imported data FIT
file.
}
\description{
Read and transform a data FIT file from csv file format to a tidy
tibble with columns and content processed for the utility of
IO.BED project.
}
\examples{
\dontrun{
sample_data <- data_test_path("8356614998_ACTIVITY.fit")
output_dir <- withr::local_tempdir()

fit_data_csv_path <- sample_data |>
  load_fit(output_dir) |>
  fit2datacsv()

parse_fit_data(fit_data_csv_path)
}

}
